+++ client/src/pages/Profile/components/Image.js
@@ -1,0 +1,48 @@
+
+export default function Image() {
+  return 
+      /* <Dropzone
+                onDrop={handleDrop}
+                noKeyboard
+                noClick={imageChanged}
+                style={{ width: "250px", height: "250px" }}
+              >
+                {({ getRootProps, getInputProps, isDragActive }) => (
+                  <div {...getRootProps()}>
+                    <AvatarEditor
+                      ref={setEditorRef}
+                      image={
+                        imagePreview ||
+                        userData.image ||
+                        "https://via.placeholder.com/200"
+                      }
+                      width={200}
+                      height={200}
+                      border={0}
+                      borderRadius={100}
+                      color={[255, 255, 255, 0]} // RGBA
+                      scale={1}
+                      rotate={0}
+                    />
+                    <input {...getInputProps()} name="pic" />
+                    {isDragActive ? <p>Drop the files here ...</p> : <p></p>}
+                  </div>
+                )}
+              </Dropzone>
+              <div id="upload-btns" className="" hidden={!imageChanged}>
+                <Button
+                  id="img-save"
+                  onClick={handleClick}
+                  className="profile-btn mx-3 btn-success"
+                >
+                  Save
+                </Button>
+                <Button
+                  id="img-cancel"
+                  onClick={handleClick}
+                  className="profile-btn mx-3 btn-danger"
+                >
+                  Cancel
+                </Button>
+              </div> */
+}
