+++ client/src/pages/Profile/components/index.js
@@ -1,0 +1,4 @@
+import Image from "./Image";
+import Form from "./Form";
+import Orders from "./Orders";
+export { Image, Form, Orders };
